requirements.txt
Django==5.2.7
djangorestframework==3.15.1
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
whitenoise==6.7.0
psycopg2-binary==2.9.9  # For PostgreSQL (optional)